+++ frontend/DDMQuestionnaire/src/components/questions/QuestionTransition.vue
@@ -12,4 +12,7 @@ const props = defineProps<{
 </template>
 
 <style scoped>
+.question-text {
+  border-bottom: none;
+}
 </style>