+++ frontend/DDMQuestionnaire/src/composables/useScrollHandler.ts
@@ -1,10 +1,9 @@
-import {onBeforeUnmount, onMounted, Ref, ref} from 'vue';
-import {getHeightOfLastQuestionTextBefore} from "@questionnaire/utils/scrollFunctions";
+import { onBeforeUnmount, onMounted, ref } from 'vue';
 
 /**
  * Composable handling automated scrolling behaviour.
  */
-export function useScrollHandler(rootElement: Ref<HTMLElement | null>) {
+export function useScrollHandler() {
   const highlightUntilWidth: number = 768;
   const shouldHighlightOnScroll = ref<boolean>(window.innerWidth <= highlightUntilWidth);
 
@@ -22,8 +21,6 @@ export function useScrollHandler(rootElement: Ref<HTMLElement | null>) {
    */
   function registerScrollEvents(): void {
     window.addEventListener('resize', updateShouldHighlightOnScroll);
-    window.addEventListener('scroll', updateMostVisibleRow);
-    window.addEventListener('load', updateMostVisibleRow);
   }
 
   /**
@@ -31,8 +28,6 @@ export function useScrollHandler(rootElement: Ref<HTMLElement | null>) {
    */
   function removeScrollEvents(): void {
     window.removeEventListener('resize', () => {});
-    window.removeEventListener('scroll', updateMostVisibleRow);
-    window.removeEventListener('load', updateMostVisibleRow);
   }
 
   /**
@@ -40,42 +35,6 @@ export function useScrollHandler(rootElement: Ref<HTMLElement | null>) {
    */
   function updateShouldHighlightOnScroll(): void {
     shouldHighlightOnScroll.value = window.innerWidth <= highlightUntilWidth;
-  }
-
-  /**
-   * Highlights the most visible `.response-row` element in the viewport by setting its opacity to `1`,
-   * while dimming all others by setting their opacity to `0.3`.
-   *
-   * If `shouldHighlightOnScroll` is false, all rows remain fully visible.
-   */
-  function updateMostVisibleRow(): void {
-    const root = rootElement.value;
-    const responseRows = Array.from(root.querySelectorAll<HTMLElement>('.response-row'));
-    let mostVisibleRow: HTMLElement | null = null;
-    let minDistance = Infinity;
-
-    if (!shouldHighlightOnScroll.value) {
-      responseRows.forEach(row => (row.style.opacity = "1"));
-      return;
-    }
-
-    responseRows.forEach(row => {
-      const rect = row.getBoundingClientRect();
-      const questionHeight = getHeightOfLastQuestionTextBefore(row);
-      const distance = Math.abs(rect.top);
-
-      const isVisible = rect.bottom > 0 && rect.top < window.innerHeight && rect.top >= (questionHeight * 0.25);
-
-      if (isVisible && distance < minDistance) {
-        minDistance = distance;
-        mostVisibleRow = row;
-      }
-    });
-
-    responseRows.forEach(row => (row.style.opacity = "0.3"));
-    if (mostVisibleRow) {
-      mostVisibleRow.style.opacity = "1";
-    }
   }
 
   /**